Auth in hikes and rides modified

+++ hikes_api_service/routers/hikes.py
@@ -73,16 +73,24 @@ def get_all_hikes(
 @router.post("/userhikes/")
 def sign_up_for_hike(
     HikeUser: HikeUser,
+    account_data: dict = Depends(authenticator.get_current_account_data),
     repo: UserHikesRepository = Depends()) -> Union[HikeUser, Error]:
     return repo.sign_up(HikeUser)
 
-@router.delete("/userhikes/{user_id}/{hike_id}", response_model=bool)
+@router.delete("/userhikes/{user_id}/{hike_id}", response_model=Union[bool, dict])
 def unjoin_hike(
     hike_id: int,
     user_id: int,
+    response: Response,
+    account_data: dict = Depends(authenticator.get_current_account_data),
     repo: UserHikesRepository = Depends(),
 ) -> bool:
-    return repo.delete(hike_id, user_id)
+    account_id = account_data["user_id"]
+    if account_id == user_id:
+        return repo.delete(hike_id, user_id)
+    else:
+        response.status = 401
+        return {"message": "Unauthorized: ID doesn't match current user"}
 
 @router.get("/users/{user_id}/hikes", response_model=Union[List[HikeOut], Error])
 def get_user_hikes(
